Area query: change area query template so that leaf_found_op gets run only when an overlapping leaf box is found.
Previously, leaf_found_op got run whenever a candidate leaf box
for overlap was found, and leaf_found_op was responsible for
checking overlap.

This change allows leaf_found_op to assume that all the leaves
given to it are part of the area query, which simplifies the
task inside leaf_found_op.